DreamHack/Level5/ShareProfile Revenge/deploy/requirements.txt
flask==3.0.3
selenium==4.15.2
PyJWT==2.10.1